+++ opencascade-hs/cpp/hs_types.h
@@ -105,6 +105,7 @@ typedef int HLRBRep_TypeOfResultingEdge;
 typedef void BOPAlgo_Builder;
 typedef void BOPAlgo_BOP;
 typedef int BOPAlgo_Operation;
+typedef void GCPnts_AbscissaPoint;
 
 #define Handle(X) void
 #define ARRAY_1(X) void